--- .vscode/settings.json
@@ -1,7 +1,0 @@
-{
-    "editor.formatOnType": true,
-    "files.associations": {
-        "*.tpp": "cpp"
-    },
-    "terminal.integrated.shell.windows": "cmd.exe"
-}
