--- .vscode/tasks.json
@@ -1,26 +1,0 @@
-{
-    "version": "2.0.0",
-    "tasks": [
-        {
-            "label": "Build C++ project",
-            "type": "shell",
-            "group": {
-                "kind": "build",
-                "isDefault": true
-            },
-            "command": "make",
-        },
-        {
-            "label": "Build & run C++ project",
-            "type": "shell",
-            "group": {
-                "kind": "test",
-                "isDefault": true
-            },
-            "command": "make",
-            "args": [
-                "run"
-            ]
-        }
-    ]
-}
