+++ CMakeLists.txt
@@ -10,22 +10,26 @@ SET(TARGET "Geothermal")
 # the one step-X.cc file, but as you expand your project you may wish
 # to add other source files as well. If your project becomes much larger,
 # you may want to either replace the following statement by something like
-#  FILE(GLOB_RECURSE TARGET_SRC  "source/*.cc")
-#  FILE(GLOB_RECURSE TARGET_INC  "include/*.h")
-#  SET(TARGET_SRC ${TARGET_SRC}  ${TARGET_INC})
+FILE(GLOB_RECURSE TARGET_SRC  "src/*.cc")
+FILE(GLOB_RECURSE TARGET_INC  "include/*.h")
+SET(TARGET_SRC ${TARGET_SRC}  ${TARGET_INC})
 # or switch altogether to the large project CMakeLists.txt file discussed
 # in the "CMake in user projects" page accessible from the "User info"
-# page of the documentation.
+# # page of the documentation.
 SET(TARGET_SRC
-  ${PROJECT_SOURCE_DIR}/src/${TARGET}.cc
+  ${PROJECT_SOURCE_DIR}/src
   )
+SET(TARGET_INC
+  ${PROJECT_SOURCE_DIR}/include
+  )
+
 
 # Usually, you will not need to modify anything beyond this point...
 
 CMAKE_MINIMUM_REQUIRED(VERSION 2.8.8)
 
 # Set include directories
-include_directories (external)
+include_directories (include)
 
 FIND_PACKAGE(deal.II 8.5.0 QUIET
   HINTS ${deal.II_DIR} ${DEAL_II_DIR} ../ ../../ $ENV{DEAL_II_DIR}
